Removed extraneous and commented-out code, fixed file structure, and some styling
