+++ lib/create-new-cocktail-list.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const client = require('../lib/client');
+const client = require('./client');
 
 function createNewCocktailList(request, response) {
     let sql2 = "INSERT INTO cocktail_lists (list_name, summary, user_id) VALUES ($1, $2, $3)"
