+++ lib/home.js
@@ -1,6 +1,5 @@
 'use strict';
 
-
 function renderHome(request, response) {
   response.render('home', { user: global.username });
 }
