+++ lib/recipe-book.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const client = require('../lib/client');
-
+const client = require('./client');
 
 function renderRecipeBookPage(request, response) {
     let sql = "SELECT * FROM cocktail_lists WHERE user_id=$1 AND list_name=$2;"
