+++ lib/search.js
