+++ public/styles/layout.css
@@ -2,10 +2,8 @@ body {
 	height: 100%;
 	font-family: 'Lato', sans-serif;
 	background-color: #fff2ef56;
-	/* background-image: linear-gradient(to bottom, rgb(255, 251, 251), #5f1600a6 100%); */
 	background-attachment: fixed;
 	text-align: center;
-	/* text-shadow: 0 .05rem .1rem rgba(0,0,0,.5); */
   }
 
 .container {
