+++ public/styles/modules.css
@@ -1,17 +1,3 @@
-/* .square {
-    border: 1px solid black;
-    margin: 30px;
-    display: grid;
-    grid-template-columns: minmax(300px, 70%);
-    justify-items: center;
-    align-items: center;
-    justify-content: center;
-}
-
-.editDetails {
-    display: none;
-} */
-
 #header {
     display: -webkit-flex;
     display: flex;
@@ -26,8 +12,8 @@
     top: 0;
     width: 375px;
 }
-/* login */
 
+/* login */
 view{
     height: 100%;
 }
@@ -57,7 +43,7 @@ main {
     font-size: 20pt;
 }
 
-/* NAVBAR/HABURGER RULES */
+/* NAVBAR/HAMBURGER RULES */
 #nav ul li a {
     display: block;
     text-decoration: none;
@@ -139,9 +125,7 @@ main {
     font-size: 20pt;
     opacity: 60%;
 }
-#button-div-home {
 
-}
   #home-main{
     height: 1000px;
   }
@@ -195,4 +179,3 @@ main {
 #search-results-title {
     padding-top: 20px;
 }
-
