+++ server.js
@@ -15,13 +15,13 @@ app.use(express.urlencoded({ extended: true, }));
 app.use(methodOverride('_method'));
 
 
-const renderHome = require('./js/home');
-const searchCocktails  = require('./js/search')
-const cocktailHandler = require('./js/cocktailHandler.js')
-const recipeBook = require('./js/recipe-book')
-const loginHandler = require('./js/loginHandler.js')
-const viewAllRecipeBooks = require('./js/all-recipe-books.js')
-const createNewCocktailList = require('./js/create-new-cocktail-list.js')
+const renderHome = require('./lib/home');
+const searchCocktails  = require('./lib/search')
+const cocktailHandler = require('./lib/cocktailHandler.js')
+const recipeBook = require('./lib/recipe-book')
+const loginHandler = require('./lib/loginHandler.js')
+const viewAllRecipeBooks = require('./lib/all-recipe-books.js')
+const createNewCocktailList = require('./lib/create-new-cocktail-list.js')
 
 
 // ROUTES
@@ -38,10 +38,6 @@ app.delete('/delete/deletecocktail', cocktailHandler.deleteCocktail);
 app.get('/about-devs', (request, response) => {response.render('about-devs')})
 app.post('/newCocktailList', createNewCocktailList);
 app.post('/search/cocktails/extra', cocktailHandler.filterExtraIngredients);
-
-// let base = ["Light rum", "Applejack", "Gin", "Dark rum", "Sweet Vermouth", "Strawberry schnapps", "Scotch", "Apricot brandy", "Triple sec", "Southern Comfort", "Orange bitters", "Brandy", "Lemon vodka", "Blended whiskey", "Dry Vermouth", "Amaretto", "Champagne", "Coffee liqueur", "Bourbon", "Tequila", "Vodka", "Añejo rum", "Bitters", "Kahlua", "Dubonnet Rouge", "Irish whiskey", "Apple brandy", "Cherry brandy", "Creme de Cacao", "Port", "Coffee brandy", "Red wine", "Rum", "Ricard", "Sherry", "Cognac", "Sloe gin", "Galliano", "Peach Vodka", "Ouzo", "Spiced rum", "Angelica root", "Johnnie Walker", "Everclear", "Firewater", "Lager", "Whiskey", "Absolut Citron", "Pisco", "Irish cream", "Ale", "Chocolate liqueur", "Midori melon liqueur", "Sambuca", "Blackberry brandy", "Peppermint schnapps", "Creme de Cassis", "Jack Daniels"];
-
-// let ingredients = ["Lime juice", "demerara Sugar", "Sugar", "Tea", "Grenadine", "Grapefruit juice", "Carbonated water", "Apple juice", "Pineapple juice", "Lemon juice", "Sugar syrup", "Milk", "Strawberries", "Chocolate syrup", "Yoghurt", "Mango", "Ginger", "Lime", "Cantaloupe", "Berries", "Grapes", "Kiwi", "Tomato juice", "Cocoa powder", "Chocolate", "Heavy cream", "Coffee", "Water", "Espresso", "Orange", "Cranberries", "Apple cider", "Cranberry juice", "Egg yolk", "Egg", "Grape juice", "Peach nectar", "Lemon", "Lemonade", "Cider", "Sprite", "7-Up"]
 
 client.connect()
   .then(() => {
