+++ views/about-devs.ejs
@@ -4,7 +4,7 @@
 
 <body>
   <%- include('./partials/header') %>
-  <container>
+  <main>
     <div>
       HI I AM KAI
     </div>
@@ -14,7 +14,7 @@
     <div>
         HI I AM SAM
     </div>
-  </container>
+  </main>
 </body>
 
 </html>
