+++ views/database/recipe-book.ejs
@@ -10,9 +10,6 @@
       View recipes for drinks you've saved, create recipe books
       for special occassions, and even update your current recipes based on personal taste!
     </h4>
-    <!-- <form action="<%= '/newCocktailList' %>" method="get">
-    <button type="submit">Create new Cocktail List</button>
-  </form> -->
     <% cocktailArray.forEach(cocktail => { %>
     <ul>
       <img src=<%= cocktail.image_url %>>
