+++ views/partials/head.ejs
@@ -1,5 +1,3 @@
-<!-- partial -->
-
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
