--- views/search/search-results-name.ejs
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<%- include('../partials/head') %>
-
-<body>
-
-</body>
-
-</html>
