+++ views/search/search.ejs
@@ -41,6 +41,18 @@
           <button class="button" type="submit">Rum</button>
         </form>
       </section>
+      <section class="base-selector-tile">
+        <form action='/search/cocktails' method='post'>
+          <input type="hidden" name="search" value="Brandy" required>
+          <button class="button" type="submit">Rum</button>
+        </form>
+      </section>
+      <section class="base-selector-tile">
+        <form action='/search/cocktails' method='post'>
+          <input type="hidden" name="search" value="Bourbon" required>
+          <button class="button" type="submit">Rum</button>
+        </form>
+      </section>
 
     </div>
   </main>
